requirements.txt
et-xmlfile==1.1.0
numpy==1.24.4
openpyxl==3.1.2
pandas==1.5.3
polars==0.17.3
pyarrow==13.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0
six==1.16.0
typing-extensions==4.12.2
xlsx2csv==0.8.3
XlsxWriter==3.2.0
